backend/requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python
scikit-learn
pillow
